💡 refactor : id 타입을 int -> Long 으로 변경

+++ BackEnd/SOA/src/main/java/com/moham/coursemores/dto/course/CourseCreateReqDto.java
@@ -17,5 +17,5 @@ public class CourseCreateReqDto {
     private boolean visited;
     private List<LocationCreateReqDto> locationList;
     private List<String> hashtagList;
-    private List<Integer> themeIdList;
+    private List<Long> themeIdList;
 }
