+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/CourseLikeRepository.java
@@ -4,9 +4,9 @@ import com.moham.coursemores.domain.CourseLike;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseLikeRepository extends JpaRepository<CourseLike, Integer> {
+public interface CourseLikeRepository extends JpaRepository<CourseLike, Long> {
 
-    Optional<CourseLike> findByUserIdAndCourseId(int userId, int courseId);
+    Optional<CourseLike> findByUserIdAndCourseId(Long userId, Long courseId);
 
 
 }
