+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/CourseLocationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CourseLocationRepository extends JpaRepository<CourseLocation, Integer> {
-    List<CourseLocation> findByCourseId(int courseId);
+public interface CourseLocationRepository extends JpaRepository<CourseLocation, Long> {
+    List<CourseLocation> findByCourseId(Long courseId);
 }
