+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/ThemeRepository.java
@@ -3,5 +3,5 @@ package com.moham.coursemores.repository;
 import com.moham.coursemores.domain.Theme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemeRepository extends JpaRepository<Theme, Integer> {
+public interface ThemeRepository extends JpaRepository<Theme, Long> {
 }
