+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/UserRepository.java
@@ -4,7 +4,7 @@ import com.moham.coursemores.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByIdAndDeleteTimeIsNull(int userId);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByIdAndDeleteTimeIsNull(Long userId);
     Optional<User> findByProviderAndProviderId(String provider, String providerId);
 }
