+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/querydsl/CourseCustomRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface CourseCustomRepository {
-    Page<Course> searchAll(String word, int regionId, List<Integer> themeIds, Pageable pageable);
+    Page<Course> searchAll(String word, Long regionId, List<Long> themeIds, Pageable pageable);
 }
