+++ BackEnd/SOA/src/main/java/com/moham/coursemores/service/InterestService.java
@@ -4,13 +4,8 @@ import com.moham.coursemores.dto.interest.InterestCourseResDto;
 import java.util.List;
 
 public interface InterestService {
-
-    List<InterestCourseResDto> getUserInterestCourseList(int userId);
-
-    boolean checkInterest(int userId, int courseId);
-
-    void addInterestCourse(int userId, int courseId);
-
-    void deleteInterestCourse(int userId, int courseId);
-
+    List<InterestCourseResDto> getUserInterestCourseList(Long userId);
+    boolean checkInterest(Long userId, Long courseId);
+    void addInterestCourse(Long userId, Long courseId);
+    void deleteInterestCourse(Long userId, Long courseId);
 }
