+++ BackEnd/SOA/src/main/java/com/moham/coursemores/service/ProfileService.java
@@ -4,11 +4,8 @@ import com.moham.coursemores.dto.profile.UserInfoResDto;
 import com.moham.coursemores.dto.profile.UserInfoUpdateReqDto;
 
 public interface ProfileService {
-    UserInfoResDto getMyProfile(int userId);
-    void alramSetting(int userId);
-
-    void updateUserInfo(int userId, UserInfoUpdateReqDto userInfoUpdateReqDto);
-    void deleteUser(int userId);
-
-
+    UserInfoResDto getMyProfile(Long userId);
+    void alramSetting(Long userId);
+    void updateUserInfo(Long userId, UserInfoUpdateReqDto userInfoUpdateReqDto);
+    void deleteUser(Long userId);
 }
